chore(ui): extracts collection and global component maps into standalone files
