+++ packages/ui/src/providers/ComponentMap/buildComponentMap/actions.tsx
