+++ packages/ui/src/utilities/buildComponentMap.ts
@@ -1,3 +1,3 @@
+export { mapFields } from '../providers/ComponentMap/buildComponentMap/fields.js'
 export { buildComponentMap } from '../providers/ComponentMap/buildComponentMap/index.js'
-export { mapFields } from '../providers/ComponentMap/buildComponentMap/mapFields.js'
 export type * from '../providers/ComponentMap/buildComponentMap/types.js'
